premature over-optimization is the root of all evil. seriously.
EEPROMex turned out to have a neat set of features but the goal is to have a working
prototype, not an ideally optimized code.

i switched to Arduino's stock EEPROM.h library, which offers the EERPOM[] object.
it offers the physical EERPOM to be read/written much like any other array of e.g. characters.
it is good enough.

once the QA board is ready, one can go an optimize this code....